Fixed webpack and "setting" webview. Also fixed config settings name.
