+++ .vscodeignore
@@ -1,9 +1,16 @@
-.vscode
-node_modules
+.vscode/**
+.vscode-test/**
+out/**
+node_modules/**
 src/**
-package-json.lock
-tsconfig.json
+.gitignore
+.yarnrc
 webpack.config.js
+vsc-extension-quickstart.md
+**/tsconfig.json
+**/.eslintrc.json
+**/*.map
+**/*.ts
 *.leo
 leoserver.py
-.gitignore
+bufferutil
