--- find-panel/main.css
@@ -1,65 +1,0 @@
-div.row {
-  overflow-x: hidden;
-  white-space: nowrap;
-}
-
-div.col {
-  vertical-align: top;
-  width: 50%;
-  max-width: 9.5em;
-  display: inline-block;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-div.col-nav {
-  width: 50%;
-  display: inline-block;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-.mt-0 {
-  margin-top: 0;
-}
-
-.mt-4 {
-  margin-top: 4px;
-}
-
-.mb-0 {
-  margin-bottom: 0;
-}
-
-.mb-4 {
-  margin-bottom: 4px;
-}
-
-.mb-6 {
-  margin-bottom: 6px;
-}
-
-.input-holder {
-  padding: 0;
-  margin-right: 6px;
-}
-
-.align-right {
-  text-align: right;
-}
-
-.float-left {
-  float: left;
-}
-
-.float-right {
-  float: right;
-}
-
-.no-overflow {
-  text-overflow: unset !important;
-}
-
-.label-fix {
-  vertical-align: top;
-  line-height: 18px;
-}
