+++ find-panel/style.css
@@ -1,3 +1,39 @@
+html {
+  box-sizing: border-box;
+  font-size: 13px;
+  overflow: hidden;
+}
+
+html:hover {
+  overflow: overlay;
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+
+body,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+ol,
+ul {
+  margin: 0;
+  padding: 0;
+  font-weight: normal;
+}
+
+img {
+  max-width: 100%;
+  height: auto;
+}
+
 :root {
   --input-padding-vertical: 6px;
   --input-padding-horizontal: 4px;
@@ -30,8 +66,8 @@ ul {
   padding-left: var(--container-padding);
 }
 
-body > *,
-form > * {
+body>*,
+form>* {
   margin-block-start: var(--input-margin-vertical);
   margin-block-end: var(--input-margin-vertical);
 }
@@ -93,13 +129,16 @@ select {
   position: relative;
   border-radius: 2px;
 }
+
 select:disabled {
   opacity: 0.5;
   cursor: default;
 }
+
 select option {
   background-color: var(--vscode-dropdown-listBackground);
 }
+
 select option:focus {
   background-color: var(--vscode-input-background);
   outline-color: var(--vscode-input-border);
@@ -124,12 +163,15 @@ input[type="checkbox"],
 input[type="radio"] {
   cursor: pointer;
 }
-input[type="checkbox"]:focus + label {
+
+input[type="checkbox"]:focus+label {
   border: 1px solid var(--vscode-button-background);
 }
-input[type="radio"]:focus + label {
+
+input[type="radio"]:focus+label {
   border: 1px solid var(--vscode-button-background);
 }
+
 input:disabled {
   opacity: 0.5;
   cursor: default;
@@ -146,8 +188,8 @@ u {
   text-underline-offset: 0.04em;
 }
 
-body > div > div > label:focus,
-body > div > div > label:active {
+body>div>div>label:focus,
+body>div>div>label:active {
   outline-color: none !important;
   border: none !important;
 }
@@ -166,4 +208,71 @@ body > div > div > label:active {
 
 .comment {
   display: none;
+}
+
+div.row {
+  overflow-x: hidden;
+  white-space: nowrap;
+}
+
+div.col {
+  vertical-align: top;
+  width: 50%;
+  max-width: 9.5em;
+  display: inline-block;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
+
+div.col-nav {
+  width: 50%;
+  display: inline-block;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
+
+.mt-0 {
+  margin-top: 0;
+}
+
+.mt-4 {
+  margin-top: 4px;
+}
+
+.mb-0 {
+  margin-bottom: 0;
+}
+
+.mb-4 {
+  margin-bottom: 4px;
+}
+
+.mb-6 {
+  margin-bottom: 6px;
+}
+
+.input-holder {
+  padding: 0;
+  margin-right: 6px;
+}
+
+.align-right {
+  text-align: right;
+}
+
+.float-left {
+  float: left;
+}
+
+.float-right {
+  float: right;
+}
+
+.no-overflow {
+  text-overflow: unset !important;
+}
+
+.label-fix {
+  vertical-align: top;
+  line-height: 18px;
 }
