+++ package.json
@@ -20,7 +20,7 @@
   },
   "homepage": "https://github.com/boltex/leointeg#readme",
   "engines": {
-    "vscode": "^1.77.0"
+    "vscode": "^1.81.0"
   },
   "keywords": [
     "leo",
@@ -96,19 +96,19 @@
           "type": "boolean",
           "description": "Specifies whether to keep focus in the outline when opening a body pane on the side"
         },
-        "leoIntegration.leoGoAnywhereShortcut": {
+        "leoIntegration.goAnywhereShortcut": {
           "scope": "application",
           "default": true,
           "type": "boolean",
           "description": "Enables Leo's 'Go Anywhere' to replace 'Go to file' with the Ctrl+P keyboard shortcut when focus in in outline or body"
         },
-        "leoIntegration.leoCollapseAllShortcut": {
+        "leoIntegration.collapseAllShortcut": {
           "scope": "application",
           "default": true,
           "type": "boolean",
           "description": "Enables the Alt+'-' keyboard shortcut for collapsing all folders in the Explorer View"
         },
-        "leoIntegration.leoActivityViewShortcut": {
+        "leoIntegration.activityViewShortcut": {
           "scope": "application",
           "default": true,
           "type": "boolean",
@@ -3731,7 +3731,7 @@
         "command": "leointeg.goAnywhere",
         "key": "ctrl+p",
         "mac": "cmd+p",
-        "when": "config.leoIntegration.leoGoAnywhereShortcut && leoTreeOpened && sideBarFocus && !editorFocus && !editorTextFocus && focusedView =~ /^leoIntegration|^leoFindPanel|^leoDocuments|^leoGoto|^leoButtons/ || config.leoIntegration.leoGoAnywhereShortcut && leoTreeOpened && editorTextFocus && resourceScheme == leo"
+        "when": "config.leoIntegration.goAnywhereShortcut && leoTreeOpened && sideBarFocus && !editorFocus && !editorTextFocus && focusedView =~ /^leoIntegration|^leoFindPanel|^leoDocuments|^leoGoto|^leoButtons/ || config.leoIntegration.goAnywhereShortcut && leoTreeOpened && editorTextFocus && resourceScheme == leo"
       },
       {
         "command": "leointeg.startSearch",
@@ -3857,12 +3857,12 @@
       {
         "command": "workbench.files.action.collapseExplorerFolders",
         "key": "alt+-",
-        "when": "config.leoIntegration.leoCollapseAllShortcut && !inSearchEditor && !sideBarFocus && resourceScheme == untitled || config.leoIntegration.leoCollapseAllShortcut && !inSearchEditor && !sideBarFocus && resourceScheme == file || config.leoIntegration.leoCollapseAllShortcut && !inSearchEditor && explorerViewletFocus"
+        "when": "config.leoIntegration.collapseAllShortcut && !inSearchEditor && !sideBarFocus && resourceScheme == untitled || config.leoIntegration.collapseAllShortcut && !inSearchEditor && !sideBarFocus && resourceScheme == file || config.leoIntegration.collapseAllShortcut && !inSearchEditor && explorerViewletFocus"
       },
       {
         "command": "workbench.view.extension.leoIntegrationView",
         "key": "ctrl+shift+l",
-        "when": "config.leoIntegration.leoActivityViewShortcut && !editorHasSelection && !inSearchEditor && focusedView != workbench.view.search"
+        "when": "config.leoIntegration.activityViewShortcut && !editorHasSelection && !inSearchEditor && focusedView != workbench.view.search"
       }
     ],
     "resourceLabelFormatters": [
@@ -3877,39 +3877,33 @@
     ]
   },
   "scripts": {
-    "vscode:prepublish": "webpack --mode production",
-    "webpack": "webpack --mode development",
-    "webpack-dev": "webpack --mode development --watch",
-    "test-compile": "tsc -p ./",
-    "lint": "tslint -p ./"
+    "vscode:prepublish": "npm run package",
+    "compile": "webpack",
+    "watch": "webpack --watch",
+    "package": "webpack --mode production --devtool hidden-source-map",
+    "compile-tests": "tsc -p . --outDir out",
+    "watch-tests": "tsc -p . -w --outDir out",
+    "pretest": "npm run compile-tests && npm run compile && npm run lint",
+    "lint": "eslint src --ext ts",
+    "test": "node ./out/test/runTest.js"
   },
   "devDependencies": {
-    "@types/glob": "^7.1.3",
+    "@types/glob": "^8.1.0",
     "@types/lodash": "^4.14.177",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^14.11.2",
-    "@types/vscode": "^1.68.0",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "20.2.5",
+    "@types/vscode": "^1.81.0",
     "@types/ws": "^7.2.6",
-    "clean-webpack-plugin": "^3.0.0",
-    "csp-html-webpack-plugin": "^4.0.0",
-    "css-loader": "^3.5.3",
-    "eslint": "^7.9.0",
-    "fork-ts-checker-webpack-plugin": "^4.1.6",
-    "glob": "^7.1.6",
-    "html-loader": "^1.1.0",
-    "html-webpack-exclude-assets-plugin": "0.0.7",
-    "html-webpack-plugin": "^3.2.0",
-    "imagemin-webpack-plugin": "^2.4.2",
-    "mini-css-extract-plugin": "^0.9.0",
-    "mocha": "^7.2.0",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^8.0.2",
-    "ts-loader": "^7.0.5",
-    "tslint": "^6.1.3",
-    "typescript": "^3.9.7",
-    "vscode-test": "^1.4.0",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "eslint": "^8.41.0",
+    "glob": "^8.1.0",
+    "mocha": "^10.2.0",
+    "ts-loader": "^9.4.3",
+    "typescript": "^5.1.3",
+    "webpack": "^5.85.0",
+    "webpack-cli": "^5.1.1",
+    "@vscode/test-electron": "^2.3.2"
   },
   "dependencies": {
     "bufferutil": "^4.0.1",
