--- tsconfig.webviews.json
@@ -1,19 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"lib": [
-			"dom",
-			"dom.iterable",
-			"es2018"
-		],
-		"outDir": "dist/webviews",
-		"skipLibCheck": true
-	},
-	"include": [
-		"src/webviews/settingsPanel/**/*"
-	],
-	"exclude": [
-		"node_modules",
-		"test"
-	]
-}
